requirements.txt
streamlit
numpy
Pillow
faiss-cpu
torch
git+https://github.com/openai/CLIP.git